+++ src/Identity.ts
@@ -336,9 +336,6 @@ export const Comonad: Comonad1<URI> = {
  */
 export const ChainRec: ChainRec1<URI> = {
   URI,
-  map: map_,
-  ap: ap_,
-  chain: chain_,
   chainRec: chainRec_
 }