+++ src/Monad.ts
@@ -19,50 +19,63 @@ import {
   Applicative2,
   Applicative2C,
   Applicative3,
-  Applicative4,
-  Applicative3C
+  Applicative3C,
+  Applicative4
 } from './Applicative'
-import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain4, Chain3C } from './Chain'
-import { URIS, URIS2, URIS3, URIS4 } from './HKT'
+import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad<F> extends Applicative<F>, Chain<F> {}
+export interface Monad<F> extends Applicative<F> {
+  readonly chain: <A, B>(fa: HKT<F, A>, f: (a: A) => HKT<F, B>) => HKT<F, B>
+}
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad1<F extends URIS> extends Applicative1<F>, Chain1<F> {}
+export interface Monad1<M extends URIS> extends Applicative1<M> {
+  readonly chain: <A, B>(fa: Kind<M, A>, f: (a: A) => Kind<M, B>) => Kind<M, B>
+}
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad2<M extends URIS2> extends Applicative2<M>, Chain2<M> {}
+export interface Monad2<M extends URIS2> extends Applicative2<M> {
+  readonly chain: <E, A, B>(fa: Kind2<M, E, A>, f: (a: A) => Kind2<M, E, B>) => Kind2<M, E, B>
+}
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad2C<M extends URIS2, L> extends Applicative2C<M, L>, Chain2C<M, L> {}
+export interface Monad2C<M extends URIS2, E> extends Applicative2C<M, E> {
+  readonly chain: <A, B>(fa: Kind2<M, E, A>, f: (a: A) => Kind2<M, E, B>) => Kind2<M, E, B>
+}
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad3<M extends URIS3> extends Applicative3<M>, Chain3<M> {}
+export interface Monad3<M extends URIS3> extends Applicative3<M> {
+  readonly chain: <R, E, A, B>(fa: Kind3<M, R, E, A>, f: (a: A) => Kind3<M, R, E, B>) => Kind3<M, R, E, B>
+}
 
 /**
  * @category type classes
  * @since 2.2.0
  */
-export interface Monad3C<M extends URIS3, E> extends Applicative3C<M, E>, Chain3C<M, E> {}
+export interface Monad3C<M extends URIS3, E> extends Applicative3C<M, E> {
+  readonly chain: <R, A, B>(fa: Kind3<M, R, E, A>, f: (a: A) => Kind3<M, R, E, B>) => Kind3<M, R, E, B>
+}
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface Monad4<M extends URIS4> extends Applicative4<M>, Chain4<M> {}
+export interface Monad4<M extends URIS4> extends Applicative4<M> {
+  readonly chain: <S, R, E, A, B>(fa: Kind4<M, S, R, E, A>, f: (a: A) => Kind4<M, S, R, E, B>) => Kind4<M, S, R, E, B>
+}