+++ src/index.ts
@@ -15,7 +15,6 @@ import * as boundedJoinSemilattice from './BoundedJoinSemilattice'
 import * as boundedLattice from './BoundedLattice'
 import * as boundedMeetSemilattice from './BoundedMeetSemilattice'
 import * as category from './Category'
-import * as chain from './Chain'
 import * as chainRec from './ChainRec'
 import * as choice from './Choice'
 import * as comonad from './Comonad'
@@ -148,10 +147,6 @@ export {
    * @since 2.0.0
    */
   category,
-  /**
-   * @since 2.0.0
-   */
-  chain,
   /**
    * @since 2.0.0
    */